pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "deepagent-demo"
version = "0.1.0"
description = "Deep Agents 快速入门与功能验证（基于 langchain-ai/deepagents）"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.5.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
    "socksio>=1.0.0",
    "tavily-python>=0.7.23",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
package = true
